requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
pymongo
dnspython
backtrader
alpaca-backtrader-api
tensorflow
bs4
nltk
pandas
datetime
